feat: chat styling
